requirements.txt
fastapi[standard]
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
toml
httpx
requests
slowapi
pydantic-settings